move js into seperate modules
