+++ index.html
@@ -10,9 +10,7 @@
     <title>Dinger Poker</title>
     <link rel="icon" type="image/x-icon" href="images/poker_chip_black.png">
     <link rel="stylesheet" href="styles.css">
-    <script src="js/helpers.js" defer></script>
-    <script src="js/data.js" defer></script>
-    <script src="js/main.js"></script>
+    <script src="js/main.js" type="module"></script>
   </head>
     
   <body>
